requirements.txt
pyjwt
pytest
moto==1.3.4
sure
